# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/transaksi": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Menampilkan semua transaksi",
                "operationId": "indexTransaksi",
                "responses": {
                    "200": {
                        "description": "Daftar transaksi berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaksi"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Buat transaksi baru",
                "operationId": "storeTransaksi",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tanggal",
                                    "jumlah",
                                    "tipe"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "kategori_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-05-05"
                                    },
                                    "jumlah": {
                                        "type": "integer",
                                        "example": 100000
                                    },
                                    "tipe": {
                                        "type": "string",
                                        "enum": [
                                            "Pemasukan",
                                            "Pengeluaran"
                                        ],
                                        "example": "Pemasukan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaksi berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Transaksi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transaksi/{id}": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Menampilkan detail transaksi",
                "operationId": "showTransaksi",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dari transaksi yang ingin ditampilkan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail transaksi berhasil ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Transaksi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Mengupdate transaksi yang ada",
                "operationId": "updateTransaksi",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID transaksi yang akan diupdate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tanggal",
                                    "jumlah",
                                    "tipe"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "kategori_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-05-05"
                                    },
                                    "jumlah": {
                                        "type": "integer",
                                        "example": 100000
                                    },
                                    "tipe": {
                                        "type": "string",
                                        "enum": [
                                            "Pemasukan",
                                            "Pengeluaran"
                                        ],
                                        "example": "Pengeluaran"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaksi berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Transaksi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Menghapus transaksi",
                "operationId": "deleteTransaksi",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID transaksi yang akan dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Transaksi berhasil dihapus"
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Transaksi": {
                "required": [
                    "tanggal",
                    "jumlah",
                    "tipe"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "kategori_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "tanggal": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-05-05"
                    },
                    "jumlah": {
                        "type": "integer",
                        "example": 100000
                    },
                    "tipe": {
                        "type": "string",
                        "enum": [
                            "Pemasukan",
                            "Pengeluaran"
                        ],
                        "example": "Pemasukan"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-05T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-06T10:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}